Revert "merge amd-gfx13 into amd-arcadia (#4598)" (#4600)
This reverts commit 99a7cd7f800f284ddd1209920f52398902097733.

It should not be squashed.

+++ llvm/utils/TableGen/Common/CodeGenRegisters.h
@@ -359,7 +359,7 @@ public:
   StringRef Namespace;
   SmallVector<ValueTypeByHwMode, 4> VTs;
   RegSizeInfoByHwMode RSI;
-  uint8_t CopyCost;
+  int CopyCost;
   bool Allocatable;
   StringRef AltOrderSelect;
   uint8_t AllocationPriority;
